+++ durabletask-azuremanaged/durabletask/azuremanaged/preview/on_demand_sandbox/__init__.py
@@ -9,17 +9,17 @@ sandbox.
 
 Usage::
 
-    from durabletask.azuremanaged.preview.ondemand_sandbox import (
+    from durabletask.azuremanaged.preview.on_demand_sandbox import (
         OnDemandSandboxWorker,
         OnDemandSandboxActivitiesClient,
     )
 """
 
-from durabletask.azuremanaged.preview.ondemand_sandbox.client import OnDemandSandboxActivitiesClient
-from durabletask.azuremanaged.preview.ondemand_sandbox.client import OnDemandSandboxWorkerProfile
-from durabletask.azuremanaged.preview.ondemand_sandbox.client import OnDemandSandboxWorkerProfileOptions
-from durabletask.azuremanaged.preview.ondemand_sandbox.client import on_demand_sandbox_worker_profile
-from durabletask.azuremanaged.preview.ondemand_sandbox.worker import OnDemandSandboxWorker
+from durabletask.azuremanaged.preview.on_demand_sandbox.client import OnDemandSandboxActivitiesClient
+from durabletask.azuremanaged.preview.on_demand_sandbox.client import OnDemandSandboxWorkerProfile
+from durabletask.azuremanaged.preview.on_demand_sandbox.client import OnDemandSandboxWorkerProfileOptions
+from durabletask.azuremanaged.preview.on_demand_sandbox.client import on_demand_sandbox_worker_profile
+from durabletask.azuremanaged.preview.on_demand_sandbox.worker import OnDemandSandboxWorker
 
 __all__ = [
     "OnDemandSandboxWorker",